Refactored Field in Grid
