+++ src/main/resources/mills.txt
@@ -1,0 +1,16 @@
+(0,1,2)
+(0,9,21)
+(1,4,7)
+(2,14,23)
+(3,4,5)
+(3,10,18)
+(5,13,20)
+(6,7,8)
+(6,11,15)
+(8,12,17)
+(9,10,11)
+(12,13,14)
+(15,16,17)
+(16,19,22)
+(18,19,20)
+(21,22,23)
