+++ src/main/scala/de/htwg/se/muehle/Muehle.scala
@@ -1,11 +1,11 @@
-package de.htwg.se.muehle.model
+package de.htwg.se.muehle
 
-import de.htwg.se.muehle.model.{Field, Player}
+import de.htwg.se.muehle.model.{Grid, Player}
 import de.htwg.se.muehle.controller.Controller
 import de.htwg.se.muehle.aview.Tui
 
 object Muehle {
-  val field = new Field()
+  val field = new Grid()
   val controller = new Controller(field)
   val tui = new Tui(controller)
