+++ src/test/scala/de/htwg/se/muehle/model/GridSpec.scala
@@ -2,12 +2,12 @@ package de.htwg.se.muehle.model
 
 import org.scalatest.{Matchers, WordSpec}
 
-class FieldSpec extends WordSpec with Matchers {
+class GridSpec extends WordSpec with Matchers {
   "A Field" when { "new" should {
-      val field = Field(init = true)
+      val grid = Grid(init = true)
       "create a Grid" in {
-        field.empty_grid.length should be(field.num_fields)
-        for (i <- field.empty_grid) i.toString should be (field.empty_field)
+        grid.empty_grid.length should be(grid.num_fields)
+        for (i <- grid.empty_grid) i.toString should be (grid.empty_field)
       }
       "create a String to display" in {
         val def_out = "O-----O-----O\n" +
@@ -17,7 +17,7 @@ class FieldSpec extends WordSpec with Matchers {
           "| | O-O-O | |\n" +
           "| O---O---O |\n" +
           "O-----O-----O\n"
-        field.toString should be (def_out)
+        grid.toString should be (def_out)
       }
     }
   }
