+++ src/test/scala/de/htwg/se/muehle/model/MillSpec.scala
@@ -1,0 +1,13 @@
+package de.htwg.se.muehle.model
+
+import org.scalatest.{Matchers, WordSpec}
+
+class MillSpec extends WordSpec with Matchers {
+  "A Mill" when { "new" should {
+    val mill = Mill()
+    "get mills" in {
+     mill.mills should be(List())
+    }
+   }
+  }
+}
